Add files via upload
Welcome to “Lucky Spin”! This exciting slot machine game allows you to bet on up to 3 lines with a maximum bet of $100 and a minimum bet of $1. With 3 rows and 3 columns, you’ll be spinning the reels to see if you can hit the jackpot. The symbols and their values are defined in the symbol count and symbol value dictionaries, respectively. Deposit some money, place your bets, and let’s see if you can win big!